Update jenkinstracking.rb
I am suggesting an update to the section that filters on reports that include the 'track' resource. The original code was treating the key as 'Track', but the key in the report is actually 'Track[filename]'. I replaced the statement with an enumerable and converted the object to a string so that I could use a regex expression against it. I have tested this and it appears to be working.